+++ src/components/friends/SearchedFriendItem.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom'
 import FriendsListE from '@components/atoms/FriendsList/FriendsListE'
-import { SearchedFriendType } from '@apis/friends/FriendsType'
+import { SearchedFriendType } from '@models/apis/FriendsType'
 
 type SearchedFriendItemPropsType = {
   friendData: SearchedFriendType
