+++ src/models/apis/AuthType.ts
