+++ src/pages/friends/SearchFriends.tsx
@@ -7,7 +7,7 @@ import { Spacing } from '@components/atoms/Spacing'
 import FriendsListBItem from '@components/friends/FriendsListBItem'
 import FriendNotFound from '@components/friends/FriendNotFound'
 import { FriendsApi } from '@apis/friends/FriendsApi'
-import { FriendsType } from '@apis/friends/FriendsType'
+import { FriendsType } from '@models/apis/FriendsType'
 import styled from '@emotion/styled'
 
 const SearchFriends = () => {
